+++ src/main/java/com/example/ShopForElectronicGoods/controllers/ArticlePriceController.java
@@ -1,8 +1,10 @@
 package com.example.ShopForElectronicGoods.controllers;
 
+import com.example.ShopForElectronicGoods.Exception.ApiRequestException;
 import com.example.ShopForElectronicGoods.models.ArticlePrice;
 import com.example.ShopForElectronicGoods.services.ArticlePriceService;
 import jakarta.persistence.criteria.CriteriaBuilder;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -35,7 +37,7 @@ public class ArticlePriceController {
 
     @PostMapping("/add/{articleId}")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<ArticlePrice> addArticlePriceForArticle(@RequestBody ArticlePrice articlePrice,
+    public ResponseEntity<ArticlePrice> addArticlePriceForArticle(@Valid @RequestBody ArticlePrice articlePrice,
                                                                   @PathVariable final Integer articleId){
         ArticlePrice a = articlePriceService.addArticlePrice(articlePrice,articleId);
         return new ResponseEntity<>(a, HttpStatus.CREATED);
@@ -44,16 +46,21 @@ public class ArticlePriceController {
 
     @PutMapping("/editAP/{articlePriceId}")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<ArticlePrice> editArticlePrice(@RequestBody ArticlePrice articlePrice,
+    public ResponseEntity<ArticlePrice> editArticlePrice(@Valid @RequestBody ArticlePrice articlePrice,
                                                          @PathVariable final Integer articlePriceId){
         ArticlePrice ap = articlePriceService.editArticlePrice(articlePrice,articlePriceId);
         return ResponseEntity.ok(ap);
     }
 
-    @DeleteMapping("/articlePriceId")
+    @DeleteMapping("/{articlePriceId}")
     @PreAuthorize("hasRole('ADMIN')")
     public void deleteArticlePriceByID(@PathVariable final Integer articlePriceId){
+        ArticlePrice ap = articlePriceService.getArticlePriceById(articlePriceId);
+        if(ap == null){
+            throw new ApiRequestException("Article price by ID " + articlePriceId + " not found", HttpStatus.NOT_FOUND);
+        }
         articlePriceService.deleteArticlePriceById(articlePriceId);
+        throw new ApiRequestException("Article price by ID " + articlePriceId + " deleted successfully ", HttpStatus.OK);
     }
 
 }
